requirements.txt
# doc2text
transformers
python-docx
uvicorn
fastapi
python-multipart
# python-pythonmagick
